add rust_src_path (#21)
Rust-analyzer looks for the RUST_SRC_PATH environment variable, which is not being set. If you have rustup installed globally, it'll probably be using that RUST_SRC_PATH, which is bad because it might not be the same as the one installed by ic-nix.

The solution is to add the environment variable to ic.nix.